Update Stacks packages to their latest release.
The following Stacks dependencies are updated in this 
commit: -

[+]   Amido.Stacks.Api:  From v0.2.24 to v0.2.26
[+]  Amido.Stacks.Testing:  From v0.2.22 to v0.2.29

Packages that were updated in an earlier commit have had 
further updates, so these have also been updated: -

[+]  Serilog:  From v3.0.1 to v3.1.0
[+]  Serilog.Console:  From v4.1.0 to v5.0.0
[+]  Microsoft.AspNetCore.Authentication.JwtBearer:
       From v6.0.23 to v6.0.25
[+]  Microsoft.AspNet.WebApi.Client
       From v5.2.9 to v6.0.0
[+]  Microsoft.AspNetCore.Testing
       From v6.0.23 to v6.0.25
[+]  Microsoft.NET.Test.Sdk:  From v17.7.2 to v17.8.0
[+]  xUnit:  From v2.4.2 to v2.6.1
[+]  xUnit.runner.visualstudio:  From v2.4.5 to v2.5.3

In addition to packages that have been updated, XBehave 
has been replaced with BDDfy in the API project.  This is 
because XBehave has been deprecated and BDDfy has been 
used in the Functional Tests project.