7726 Replace Serilog from Codebase and implement Open Telemetry (#484)
* refactor: Update OpenTelemetry instrumentation in background worker

This commit updates the OpenTelemetry instrumentation in the background worker code. It adds the necessary packages and configures the logging and metrics exporters. The code now uses the `Azure.Monitor.OpenTelemetry.AspNetCore` package for exporting telemetry data to Azure Monitor. It also includes the `OpenTelemetry.Logs` and `OpenTelemetry.Metrics` packages for logging and metrics instrumentation.

* refactor: Update OpenTelemetry instrumentation in EventHub Listener

* refactor: Update OpenTelemetry instrumentation in CosmosDB worker

* refactor: Update OpenTelemetry instrumentation in ServiceBus Listener

* refactor: Update OpenTelemetry instrumentation in Simple API & CQRS

* refactor: Update OpenTelemetry instrumentation in Shared API